refactor: move psbt and tx related functions to bitcoin helper
